color datatypes fixed

+++ src/color.c
@@ -6,30 +6,30 @@
 /*   By: mbutuzov <mbutuzov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/15 18:06:34 by mbutuzov          #+#    #+#             */
-/*   Updated: 2025/01/15 18:06:37 by mbutuzov         ###   ########.fr       */
+/*   Updated: 2025/01/23 16:49:19 by mbutuzov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
 
-int32_t	get_red(int32_t color)
+int32_t	get_red(uint32_t color)
 {
-	return ((color >> 16) & 0xFF);
+	return ((int32_t)((color >> 16) & 0xFF));
 }
 
-int32_t	get_green(int32_t color)
+int32_t	get_green(uint32_t color)
 {
-	return ((color >> 8) & 0xFF);
+	return ((int32_t)((color >> 8) & 0xFF));
 }
 
-int32_t	get_blue(int32_t color)
+int32_t	get_blue(uint32_t color)
 {
-	return ((color) & 0xFF);
+	return ((int32_t)((color) & 0xFF));
 }
 
-int32_t	get_color(
-	int32_t start_color,
-	int32_t end_color,
+uint32_t	get_color(
+	uint32_t start_color,
+	uint32_t end_color,
 	int32_t i,
 	int32_t length
 )
@@ -47,5 +47,7 @@ int32_t	get_color(
 	final.red = get_red(start_color) + (i * (change.red)) / length;
 	final.green = get_green(start_color) + (i * (change.green)) / length;
 	final.blue = get_blue(start_color) + (i * (change.blue)) / length;
-	return (final.red << 16 | final.green << 8 | final.blue);
+	return (((uint32_t)final.red) << 16
+		| ((uint32_t)final.green) << 8
+		| (uint32_t)final.blue);
 }
